Ahmad Khalaf CS193 Homework 0
For my first CS193 Homework assignment, I edited the given template to make a list of my favorite things about the class. I changed most of the original text to fit the assignment's criteria. I also added a few lines of my own. The final thing I did was remove any parts of the template that weren't necessary to the assignment,